/docs/test.html is now /docs/index.html. The original file is now /docs/index_orig.html.
The hash argument of the url is used to select the page to display. The page contains both the description (as a comment) and code. I'm not sure if using # is the right thing because it makes it a little hard to switch to another page (if you edit the browser url).

Lesson files are in /docs/din10/*.lzx. default.lzx is the default lesson and is shown if the specified page does not exist.